Refactor: Move range slider parts
