+++ packages/flutter/lib/material.dart
@@ -143,6 +143,7 @@ export 'src/material/radio.dart';
 export 'src/material/radio_list_tile.dart';
 export 'src/material/radio_theme.dart';
 export 'src/material/range_slider.dart';
+export 'src/material/range_slider_parts.dart';
 export 'src/material/refresh_indicator.dart';
 export 'src/material/reorderable_list.dart';
 export 'src/material/scaffold.dart';
