+++ packages/flutter/lib/src/material/range_slider.dart
@@ -26,6 +26,7 @@ import 'package:flutter/widgets.dart';
 import 'constants.dart';
 import 'debug.dart';
 import 'material_state.dart';
+import 'range_slider_parts.dart';
 import 'slider_theme.dart';
 import 'slider_value_indicator_shape.dart';
 import 'theme.dart';
